+++ src/Logger/LoggerService.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 namespace Jtl\Connector\Core\Logger;
 
-use Composer\InstalledVersions;
+use DI\Container;
 use Jtl\Connector\Core\Exception\LoggerException;
 use Jtl\Connector\Core\Logger\Handler\ChunkedHandler;
 use Jtl\Connector\Core\Logger\Processor\RequestProcessor;
+use Jtl\Connector\Core\Logger\Processor\WarningProcessor;
+use Jtl\Connector\Core\Rpc\Warnings;
 use Monolog\Formatter\FormatterInterface;
-use Monolog\Handler\AbstractHandler;
 use Monolog\Handler\FilterHandler;
 use Monolog\Handler\FormattableHandlerInterface;
 use Monolog\Handler\HandlerInterface;
@@ -58,18 +59,20 @@ class LoggerService
      *
      * @param string     $logDir
      * @param int|string $logLevel
+     * @param Warnings   $warnings
      * @param int        $maxFiles
      *
      * @throws InvalidArgumentException
      * @throws RuntimeException
      * @throws UnexpectedValueException|\InvalidArgumentException
      */
-    public function __construct(string $logDir, int|string $logLevel, int $maxFiles = 2)
+    public function __construct(string $logDir, int|string $logLevel, Warnings $warnings, int $maxFiles = 2)
     {
         $this->logDir   = $logDir;
         $this->logLevel = $logLevel;
         $this->maxFiles = $maxFiles;
         $this
+            ->pushProcessor(new WarningProcessor($warnings))
             ->pushProcessor(new PsrLogMessageProcessor())
             ->pushProcessor(new MemoryPeakUsageProcessor())
             ->pushProcessor(new RequestProcessor());