+++ tests/src/Logger/LoggerServiceTest.php
@@ -7,6 +7,7 @@ namespace Jtl\Connector\Core\Test\Logger;
 use Composer\InstalledVersions;
 use Jtl\Connector\Core\Exception\LoggerException;
 use Jtl\Connector\Core\Logger\LoggerService;
+use Jtl\Connector\Core\Rpc\Warnings;
 use Jtl\Connector\Core\Test\TestCase;
 use Monolog\Formatter\FormatterInterface;
 use Monolog\Formatter\JsonFormatter;
@@ -225,7 +226,8 @@ class LoggerServiceTest extends TestCase
     protected function setUp(): void
     {
         parent::setUp();
+        $warnings      = new Warnings();
         $this->logDir  = \sprintf('%s/var/log', $this->connectorDir);
-        $this->factory = new LoggerService($this->logDir, LogLevel::DEBUG);
+        $this->factory = new LoggerService($this->logDir, LogLevel::DEBUG, $warnings);
     }
 }